New Templates for Food Network Redesign
The new Food Network site launch means that the template assets for that
site needed to be updated - the cartridge, recipe/product carousel,
sponsor module and tip module have been updated. They *only* work in the
new site and not the old site. The RSI module will be coming shortly as
we're reworking how that module is built.